+++ java/fury-core/src/main/java/org/apache/fury/builder/BaseObjectCodecBuilder.java
@@ -87,10 +87,10 @@ import org.apache.fury.resolver.ClassInfoHolder;
 import org.apache.fury.resolver.ClassResolver;
 import org.apache.fury.resolver.RefResolver;
 import org.apache.fury.serializer.CompatibleSerializer;
+import org.apache.fury.serializer.EnumSerializer;
 import org.apache.fury.serializer.ObjectSerializer;
 import org.apache.fury.serializer.PrimitiveSerializers.LongSerializer;
 import org.apache.fury.serializer.Serializer;
-import org.apache.fury.serializer.Serializers;
 import org.apache.fury.serializer.StringSerializer;
 import org.apache.fury.serializer.collection.AbstractCollectionSerializer;
 import org.apache.fury.serializer.collection.AbstractMapSerializer;
@@ -280,7 +280,7 @@ public abstract class BaseObjectCodecBuilder extends CodecBuilder {
         Fury.class, MemoryBuffer.class, fury.getRefResolver().getClass(), Platform.class);
     ctx.addImports(ClassInfo.class, ClassInfoHolder.class, ClassResolver.class);
     ctx.addImport(Generated.class);
-    ctx.addImports(LazyInitBeanSerializer.class, Serializers.EnumSerializer.class);
+    ctx.addImports(LazyInitBeanSerializer.class, EnumSerializer.class);
     ctx.addImports(Serializer.class, StringSerializer.class);
     ctx.addImports(ObjectSerializer.class, CompatibleSerializer.class);
     ctx.addImports(AbstractCollectionSerializer.class, AbstractMapSerializer.class);
